@@ -14,48 +14,93 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "14.2.0",
-    "@angular/cdk": "14.2.0",
-    "@angular/common": "14.2.0",
-    "@angular/compiler": "14.2.0",
-    "@angular/core": "14.2.0",
-    "@angular/flex-layout": "14.0.0-beta.40",
-    "@angular/forms": "14.2.0",
-    "@angular/material": "14.2.0",
-    "@angular/platform-browser": "14.2.0",
-    "@angular/platform-browser-dynamic": "14.2.0",
-    "@angular/router": "14.2.0",
-    "rxjs": "7.5.6",
-    "tslib": "2.4.0",
-    "zone.js": "0.11.8"
+    "@angular-devkit/core": "^19.2.13",
+    "@angular-devkit/schematics": "^19.2.13",
+    "@angular/animations": "19.2.13",
+    "@angular/cdk": "19.2.17",
+    "@angular/common": "19.2.13",
+    "@angular/compiler": "19.2.13",
+    "@angular/core": "19.2.13",
+    "@angular/flex-layout": "15.0.0-beta.42",
+    "@angular/forms": "19.2.13",
+    "@angular/material": "19.2.17",
+    "@angular/platform-browser": "19.2.13",
+    "@angular/platform-browser-dynamic": "19.2.13",
+    "@angular/router": "^19.2.13",
+    "@schematics/angular": "^19.2.13",
+    "rxjs": "7.8.2",
+    "tslib": "^2.8.1",
+    "zone.js": "0.15.1"
   },
   "devDependencies": {
-    "@angular-builders/jest": "14.0.1",
-    "@angular-builders/custom-webpack": "^14.0.1",
-    "@angular-devkit/build-angular": "14.2.1",
-    "@angular-eslint/builder": "14.0.3",
-    "@angular-eslint/eslint-plugin": "14.0.3",
-    "@angular-eslint/eslint-plugin-template": "14.0.3",
-    "@angular-eslint/schematics": "14.0.3",
-    "@angular-eslint/template-parser": "14.0.3",
-    "@angular/cli": "14.2.1",
-    "@angular/compiler-cli": "14.2.0",
-    "@briebug/jest-schematic": "4.0.0",
-    "@cypress/code-coverage": "3.10.0",
-    "@cypress/schematic": "2.0.3",
+    "@angular-builders/custom-webpack": "^19.0.1",
+    "@angular-builders/jest": "19.0.1",
+    "@angular-devkit/build-angular": "19.2.13",
+    "@angular-eslint/builder": "19.5.0",
+    "@angular-eslint/eslint-plugin": "19.5.0",
+    "@angular-eslint/eslint-plugin-template": "19.5.0",
+    "@angular-eslint/schematics": "19.5.0",
+    "@angular-eslint/template-parser": "19.5.0",
+    "@angular/cli": "19.2.13",
+    "@angular/compiler-cli": "19.2.13",
+    "@cypress/code-coverage": "^3.14.3",
+    "@cypress/schematic": "3.0.0",
     "@istanbuljs/nyc-config-typescript": "1.0.2",
     "@jsdevtools/coverage-istanbul-loader": "3.0.5",
-    "@types/jest": "28.1.3",
-    "@typescript-eslint/eslint-plugin": "5.35.1",
-    "@typescript-eslint/parser": "5.35.1",
-    "cypress": "10.4.0",
-    "eslint": "8.23.0",
-    "istanbul-lib-coverage": "3.2.0",
-    "jest": "28.1.3",
-    "jest-html-reporter": "3.6.0",
-    "jest-marbles": "3.0.3",
-    "nyc": "15.1.0",
-    "typescript": "4.7.4",
-    "webpack": "5.74.0"
+    "@types/babel__core": "^7.20.5",
+    "@types/babel__generator": "^7.6.4",
+    "@types/babel__template": "^7.4.1",
+    "@types/babel__traverse": "^7.18.0",
+    "@types/body-parser": "^1.19.2",
+    "@types/bonjour": "^3.5.13",
+    "@types/connect": "^3.4.35",
+    "@types/connect-history-api-fallback": "^1.5.4",
+    "@types/eslint": "^8.4.6",
+    "@types/eslint-scope": "^8.3.0",
+    "@types/estree": "^1.0.7",
+    "@types/express": "^4.17.22",
+    "@types/express-serve-static-core": "^4.19.6",
+    "@types/graceful-fs": "^4.1.9",
+    "@types/http-errors": "^2.0.4",
+    "@types/http-proxy": "^1.17.16",
+    "@types/istanbul-lib-coverage": "^2.0.4",
+    "@types/istanbul-lib-report": "^3.0.0",
+    "@types/istanbul-reports": "^3.0.4",
+    "@types/jest": "^29.5.14",
+    "@types/jsdom": "^21.1.7",
+    "@types/json-schema": "^7.0.15",
+    "@types/mime": "^3.0.4",
+    "@types/node": "^18.7.13",
+    "@types/node-forge": "^1.3.11",
+    "@types/qs": "^6.9.7",
+    "@types/range-parser": "^1.2.4",
+    "@types/retry": "^0.12.5",
+    "@types/send": "^0.17.4",
+    "@types/serve-index": "^1.9.4",
+    "@types/serve-static": "^1.15.7",
+    "@types/sinonjs__fake-timers": "^8.1.5",
+    "@types/sizzle": "^2.3.9",
+    "@types/sockjs": "^0.3.36",
+    "@types/stack-utils": "^2.0.3",
+    "@types/tough-cookie": "^4.0.2",
+    "@types/ws": "^8.18.1",
+    "@types/yargs": "^17.0.33",
+    "@types/yargs-parser": "^21.0.0",
+    "@types/yauzl": "^2.10.3",
+    "@typescript-eslint/eslint-plugin": "8.32.1",
+    "@typescript-eslint/parser": "8.32.1",
+    "cypress": "14.4.0",
+    "cypress-slow-down": "^1.3.1",
+    "eslint": "9.27.0",
+    "istanbul-instrumenter-loader": "^3.0.1",
+    "istanbul-lib-coverage": "3.2.2",
+    "jest": "^29.7.0",
+    "jest-html-reporter": "4.1.0",
+    "jest-marbles": "3.1.1",
+    "jest-preset-angular": "^14.5.5",
+    "nyc": "17.1.0",
+    "ts-jest": "^29.3.4",
+    "typescript": "5.8.3",
+    "webpack": "5.99.9"
   }
 }
