@@ -17,10 +17,10 @@ public class JwtUtils {
   private static final Logger logger = LoggerFactory.getLogger(JwtUtils.class);
 
   @Value("${oc.app.jwtSecret}")
-  private String jwtSecret;
+public String jwtSecret;
 
   @Value("${oc.app.jwtExpirationMs}")
-  private int jwtExpirationMs;
+public int jwtExpirationMs;
 
   public String generateJwtToken(Authentication authentication) {
 
