@@ -1,16 +1,15 @@
-import { defineConfig } from 'cypress'
+import { defineConfig } from 'cypress';
+
+// au lieu de `import…`, utilise require :
+const codeCoverageTask = require('@cypress/code-coverage/task');
 
 export default defineConfig({
-  videosFolder: 'cypress/videos',
-  screenshotsFolder: 'cypress/screenshots',
-  fixturesFolder: 'cypress/fixtures',
-  video: false,
   e2e: {
-    // We've imported your old cypress plugins here.
-    // You may want to clean this up later by importing these.
     setupNodeEvents(on, config) {
-      return require('./cypress/plugins/index.ts').default(on, config)
+      codeCoverageTask(on, config); // <— appelle directement la fonction renvoyée
+      return config;
     },
     baseUrl: 'http://localhost:4200',
+    supportFile: 'cypress/support/e2e.ts',
   },
-})
+});
