@@ -1,8 +1,8 @@
-/**
- * @type {Cypress.PluginConfig}
- */
- import * as registerCodeCoverageTasks from '@cypress/code-coverage/task';
+import * as codeCoverageTasks from '@cypress/code-coverage/task';
 
- export default (on, config) => {
-   return registerCodeCoverageTasks(on, config);
- };
+const pluginConfig: Cypress.PluginConfig = (on, config) => {
+  codeCoverageTasks.registerCodeCoverageTasks(on, config);
+  return config;
+};
+
+export default pluginConfig;
