@@ -14,6 +14,5 @@
 // ***********************************************************
 
 // When a command from ./commands is ready to use, import with `import './commands'` syntax
-// import './commands';
-
+import './commands';
 import '@cypress/code-coverage/support';
