@@ -1,5 +1,7 @@
-import 'jest-preset-angular/setup-jest';
+import { setupZoneTestEnv } from 'jest-preset-angular/setup-env/zone';
 
+setupZoneTestEnv();
+      
 /* global mocks for jsdom */
 const mock = () => {
   let storage: { [key: string]: string } = {};
