@@ -7,7 +7,8 @@ import { RegisterRequest } from '../../interfaces/registerRequest.interface';
 @Component({
   selector: 'app-register',
   templateUrl: './register.component.html',
-  styleUrls: ['./register.component.scss']
+  styleUrls: ['./register.component.scss'],
+  standalone: false
 })
 export class RegisterComponent {
 
