@@ -11,7 +11,8 @@ import { SessionApiService } from '../../services/session-api.service';
 @Component({
   selector: 'app-detail',
   templateUrl: './detail.component.html',
-  styleUrls: ['./detail.component.scss']
+  styleUrls: ['./detail.component.scss'],
+  standalone: false
 })
 export class DetailComponent implements OnInit {
   public session: Session | undefined;
