@@ -8,7 +8,8 @@ import { SessionApiService } from '../../services/session-api.service';
 @Component({
   selector: 'app-list',
   templateUrl: './list.component.html',
-  styleUrls: ['./list.component.scss']
+  styleUrls: ['./list.component.scss'],
+  standalone: false
 })
 export class ListComponent {
 
