@@ -15,5 +15,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class SessionsRoutingModule {
-}
+export class SessionsRoutingModule {}
