@@ -1,22 +1,44 @@
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
-import { expect } from '@jest/globals';
-
 import { UserService } from './user.service';
+import { expect } from '@jest/globals';
 
 describe('UserService', () => {
   let service: UserService;
+  let httpMock: HttpTestingController;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports:[
-        HttpClientModule
-      ]
+      imports: [HttpClientTestingModule]
     });
     service = TestBed.inject(UserService);
+    httpMock = TestBed.inject(HttpTestingController);
+  });
+
+  afterEach(() => {
+    httpMock.verify();
   });
 
   it('should be created', () => {
     expect(service).toBeTruthy();
+  });
+
+  it('should fetch user by id', () => {
+    const mockUser = { id: 1, firstName: 'Alyssa', lastName: 'Test' };
+    service.getById('1').subscribe(user => {
+      expect(user).toEqual(mockUser);
+    });
+    const req = httpMock.expectOne('api/user/1');
+    expect(req.request.method).toBe('GET');
+    req.flush(mockUser);
+  });
+
+  it('should delete user by id', () => {
+    service.delete('1').subscribe(response => {
+      expect(response).toEqual({});
+    });
+    const req = httpMock.expectOne('api/user/1');
+    expect(req.request.method).toBe('DELETE');
+    req.flush({});
   });
 });
